Update _config.yml (#13)
* Update _config.yml

to fix the Sympy reference

* Update 01.ipynb

Add some markdowns to align the errors and debugging section to MUDE corresponding part

* Update sympy.ipynb

please ignore the first pull request, which changed the config.yml. config is completely fine and i just forget to add the reference. after adding the reference it should be completely fine to show the reference part.