Sphinx thebe from GitHub (#39)
* Update requirements.txt

* Update _config.yml

to fix the Sympy reference

* Update 01.ipynb

Add some markdowns to align the errors and debugging section to MUDE corresponding part

* Update sympy.ipynb

please ignore the first pull request, which changed the config.yml. config is completely fine and i just forget to add the reference. after adding the reference it should be completely fine to show the reference part.

* Update _config.yml (#13)

* Update _config.yml

to fix the Sympy reference

* Update 01.ipynb

Add some markdowns to align the errors and debugging section to MUDE corresponding part

* Update sympy.ipynb

please ignore the first pull request, which changed the config.yml. config is completely fine and i just forget to add the reference. after adding the reference it should be completely fine to show the reference part.

* Update _config.yml

to make it the original

* First branch (#14)

* Update _config.yml (#17)

* Update 01.ipynb (#18)

* Fix figures (#19)

* fix figures

* fix 04

* MyST

* myst_try

* fix by MyST

* simple bug fix  (#25)

* fix bugs for book6

* fix typo in chapter 5

* matrix operator fix

wrong matrix operator in chapter 5

* matrix operator fix

* H5p test(only changed Exercise 1.2.2) (#22)

* exercise transition

* test

* update readme

* create docs subdir and add old stuff

* describe quiz convention

* apply quiz convention

* book1_exercises_updata

* remove-input

* remove-input

* remove_input

---------

Co-authored-by: Robert Lanzafame <R.C.Lanzafame@tudelft.nl>

* Chapter2  h5p (#27)

* h5p--chapter2

* add notes to h5p

* chapter1- typo fix (#28)

* Chapter5 h5p (#29)

* chapter5-h5p

* exercise2

* bug form remove

* reportbug-setup

* intro comments and filename formatting

* intro text: adjust additional resources info

* update feedback instructions

* toolbox-calculator update

* fix intro

* better name

* get sphinx-thebe from github

---------

Co-authored-by: Tom van Woudenberg <t.r.vanwoudenberg@tudelft.nl>
Co-authored-by: tshiya731 <tshiya05@gmail.com>
Co-authored-by: tshiya731 <152382463+tshiya731@users.noreply.github.com>